requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pandas>=2.2.2
google-adk==1.0.0
google-generativeai>=0.8.0